Move source code to a single file
