+++ .prettierignore
@@ -1,6 +1,3 @@
-# Artifacts
-dist
-
 # Coverage directory
 coverage
