--- postbuild.js
@@ -1,3 +1,0 @@
-import fs from "fs";
-
-fs.rmSync("./dist/types.js", { force: true });
