+++ tsconfig.tslint.json
@@ -1,4 +1,4 @@
 {
   "extends": "./tsconfig.json",
-  "exclude": ["src/tests/index.errors.ts"]
+  "exclude": ["**/*.errors.ts"]
 }
