--- tsconfig.types.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./tsconfig.build.json",
-  "compilerOptions": {
-    "declaration": true,
-    "declarationDir": "./dist",
-    "emitDeclarationOnly": true
-  },
-  "exclude": [],
-  "include": ["src/types.ts"]
-}
